1. pif_pulse에오류 수정함.
  - uint32_t 차이를 uint16_t에 저장 → 65ms 초과 시 값 절삭
  - __count가 rising 에지에서도 증가 → 유효성 판단 오류
  - SetValidRange 범위 초과 입력값 검사 없음
  - struct 이름 오타 StPifPulsData
   - Doxygen @fn 태그와 실제 함수명 불일치
  - ColSig_Init/Clear 헤더 선언 누락

+++ source/core/pif_pulse.c
@@ -105,7 +105,7 @@ void pifPulse_ResetMeasureMode(PifPulse* p_owner, uint8_t measure_mode)
 	p_owner->_measure_mode &= ~measure_mode;
 }
 
-BOOL pifPulse_SetValidRange(PifPulse* p_owner, uint8_t measure_mode, uint32_t min, uint32_t max)
+BOOL pifPulse_SetValidRange(PifPulse* p_owner, uint8_t measure_mode, uint16_t min, uint16_t max)
 {
 	int index = -1;
 
@@ -142,10 +142,10 @@ void pifPulse_ResetMeasureValue(PifPulse* p_owner)
 	p_owner->__count = 0;
 }
 
-uint16_t pifPulse_GetPeriod(PifPulse* p_owner)
+uint32_t pifPulse_GetPeriod(PifPulse* p_owner)
 {
 	uint8_t prev;
-	uint16_t value = 0;
+	uint32_t value = 0;
 
 	if (p_owner->__count < PIF_PULSE_DATA_SIZE) return 0;
 
@@ -160,10 +160,10 @@ uint16_t pifPulse_GetPeriod(PifPulse* p_owner)
 	return value;
 }
 
-uint16_t pifPulse_GetLowWidth(PifPulse* p_owner)
+uint32_t pifPulse_GetLowWidth(PifPulse* p_owner)
 {
 	uint8_t prev;
-	uint16_t value = 0;
+	uint32_t value = 0;
 
 	if (p_owner->__count < PIF_PULSE_DATA_SIZE) return 0;
 
@@ -178,9 +178,9 @@ uint16_t pifPulse_GetLowWidth(PifPulse* p_owner)
 	return value;
 }
 
-uint16_t pifPulse_GetHighWidth(PifPulse* p_owner)
+uint32_t pifPulse_GetHighWidth(PifPulse* p_owner)
 {
-	uint16_t value = 0;
+	uint32_t value = 0;
 
 	if (p_owner->__count < PIF_PULSE_DATA_SIZE) return 0;
